requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.7
annotated-types==0.6.0
anyio==4.0.0
attrs==23.2.0
bar==0.2.1
barbar==0.2.1
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fake-useragent==1.5.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.23.1
humanfriendly==10.0
idna==3.4
imageio==2.34.1
imread-from-url==0.1.3
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
magic-filter==1.0.12
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
onnx==1.16.0
onnxruntime-gpu==1.18.0
openai==1.13.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
pyreadline3==3.4.1
pyTelegramBotAPI==4.14.1
python-dateutil==2.9.0.post0
python-telegram-bot==20.5
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.3
tinydb==4.8.0
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.3
xmltodict==0.13.0
yarl==1.9.4